Updating Diameter JAIN SLEE RAs for Dictionary improvements
Update Issue #2222
Updated affected RAs, since enumerated is now being correctly handled as integer instead of long, no need to cast anymore.



git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@16928 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ resources/diameter-base/common/events/src/main/java/org/mobicents/slee/resource/diameter/base/events/AccountingMessageImpl.java
@@ -131,7 +131,7 @@ public abstract class AccountingMessageImpl extends DiameterMessageImpl implemen
   }
 
   public void setAccountingRealtimeRequired(AccountingRealtimeRequiredType accountingRealtimeRequired) {
-    addAvp(Avp.ACCOUNTING_REALTIME_REQUIRED, (long)accountingRealtimeRequired.getValue());
+    addAvp(Avp.ACCOUNTING_REALTIME_REQUIRED, accountingRealtimeRequired.getValue());
   }
 
 }
