+++ resources/diameter-base/common/events/src/main/java/org/mobicents/slee/resource/diameter/base/events/DisconnectPeerRequestImpl.java
@@ -64,7 +64,7 @@ public class DisconnectPeerRequestImpl extends DiameterMessageImpl implements Di
   }
 
   public void setDisconnectCause(DisconnectCauseType disconnectCause) {
-    addAvp(Avp.DISCONNECT_CAUSE, (long)disconnectCause.getValue());
+    addAvp(Avp.DISCONNECT_CAUSE, disconnectCause.getValue());
   }
 
 }
