+++ resources/diameter-base/common/events/src/main/java/org/mobicents/slee/resource/diameter/base/events/SessionTerminationRequestImpl.java
@@ -63,7 +63,7 @@ public class SessionTerminationRequestImpl  extends SessionTerminationMessageImp
   }
 
   public void setTerminationCause(TerminationCauseType terminationCause) {
-    addAvp(Avp.TERMINATION_CAUSE, (long)terminationCause.getValue());
+    addAvp(Avp.TERMINATION_CAUSE, terminationCause.getValue());
   }
