+++ resources/diameter-cca/common/events/src/main/java/org/mobicents/slee/resource/diameter/cca/events/avp/UsedServiceUnitAvpImpl.java
@@ -68,7 +68,7 @@ public class UsedServiceUnitAvpImpl extends ServiceUnitAvpTypeImpl implements Us
    * (net.java.slee.resource.diameter.cca.events.avp.TariffChangeUsageType)
    */
   public void setTariffChangeUsage(TariffChangeUsageType tariffChangeUsage) {
-    addAvp(CreditControlAVPCodes.Tariff_Change_Usage, (long)tariffChangeUsage.getValue());
+    addAvp(CreditControlAVPCodes.Tariff_Change_Usage, tariffChangeUsage.getValue());
   }
 
 }
