+++ resources/diameter-cx-dx/events/src/main/java/org/mobicents/slee/resource/diameter/cxdx/events/MultimediaAuthenticationRequestImpl.java
@@ -126,7 +126,7 @@ public class MultimediaAuthenticationRequestImpl extends DiameterMessageImpl imp
    * @see net.java.slee.resource.diameter.cxdx.events.MultimediaAuthenticationRequest#setAuthSessionState(net.java.slee.resource.diameter.base.events.avp.AuthSessionStateType)
    */
   public void setAuthSessionState(AuthSessionStateType authSessionState) {
-    addAvp(DiameterAvpCodes.AUTH_SESSION_STATE, (long)authSessionState.getValue());
+    addAvp(DiameterAvpCodes.AUTH_SESSION_STATE, authSessionState.getValue());
   }
 
   /* (non-Javadoc)
