+++ resources/diameter-cx-dx/events/src/main/java/org/mobicents/slee/resource/diameter/cxdx/events/RegistrationTerminationAnswerImpl.java
@@ -107,7 +107,7 @@ public class RegistrationTerminationAnswerImpl extends DiameterMessageImpl imple
    * @see net.java.slee.resource.diameter.cxdx.events.RegistrationTerminationAnswer#setAuthSessionState(net.java.slee.resource.diameter.base.events.avp.AuthSessionStateType)
    */
   public void setAuthSessionState(AuthSessionStateType authSessionState) {
-    addAvp(DiameterAvpCodes.AUTH_SESSION_STATE, (long)authSessionState.getValue());
+    addAvp(DiameterAvpCodes.AUTH_SESSION_STATE, authSessionState.getValue());
   }
 
   /* (non-Javadoc)
