+++ resources/diameter-gx/events/src/main/java/org/mobicents/slee/resource/diameter/gx/events/GxCreditControlMessageImpl.java
@@ -85,7 +85,7 @@ public abstract class GxCreditControlMessageImpl extends DiameterMessageImpl imp
      */
     @Override
     public void setCcRequestType(CcRequestType ccRequestType) throws IllegalStateException {
-        addAvp(CreditControlAVPCodes.CC_Request_Type, (long) ccRequestType.getValue());
+        addAvp(CreditControlAVPCodes.CC_Request_Type, ccRequestType.getValue());
     }
 
     /*
