+++ resources/diameter-ro/events/src/main/java/org/mobicents/slee/resource/diameter/ro/events/RoCreditControlRequestImpl.java
@@ -120,7 +120,7 @@ public class RoCreditControlRequestImpl extends RoCreditControlMessageImpl imple
    */
   @Override
   public void setTerminationCause(TerminationCauseType terminationCause) throws IllegalStateException {
-    addAvp(DiameterAvpCodes.TERMINATION_CAUSE, (long) terminationCause.getValue());
+    addAvp(DiameterAvpCodes.TERMINATION_CAUSE, terminationCause.getValue());
   }
 
   /* 
@@ -147,7 +147,7 @@ public class RoCreditControlRequestImpl extends RoCreditControlMessageImpl imple
    */
   @Override
   public void setRequestedAction(RequestedActionType requestedAction) throws IllegalStateException {
-    addAvp(CreditControlAVPCodes.Requested_Action, (long) requestedAction.getValue());
+    addAvp(CreditControlAVPCodes.Requested_Action, requestedAction.getValue());
   }
 
   /* 
@@ -174,7 +174,7 @@ public class RoCreditControlRequestImpl extends RoCreditControlMessageImpl imple
    */
   @Override
   public void setMultipleServicesIndicator(MultipleServicesIndicatorType multipleServicesIndicator) throws IllegalStateException {
-    addAvp(CreditControlAVPCodes.Multiple_Services_Indicator, (long) multipleServicesIndicator.getValue());
+    addAvp(CreditControlAVPCodes.Multiple_Services_Indicator, multipleServicesIndicator.getValue());
   }
 
   /*
